Cleanup the AEM article database logic

+++ ui/article-aem-renderer/src/main/java/org/cru/godtools/article/aem/db/TranslationRepository.kt
@@ -1,7 +1,6 @@
 package org.cru.godtools.article.aem.db
 
 import android.net.Uri
-import androidx.annotation.WorkerThread
 import androidx.room.Dao
 import androidx.room.Transaction
 import org.cru.godtools.article.aem.model.AemImport
@@ -12,25 +11,12 @@ import org.cru.godtools.article.aem.model.toTranslationRefKey
 import org.cru.godtools.model.Translation
 
 @Dao
-abstract class TranslationRepository internal constructor(private val db: ArticleRoomDatabase) {
-    @WorkerThread
-    fun find(key: Key?): TranslationRef? {
-        return key?.run { db.translationDao().find(tool, language, version) }
-    }
-
-    @WorkerThread
-    fun isProcessed(translation: Translation): Boolean {
-        return find(translation.toTranslationRefKey())?.processed ?: false
-    }
-
-    @WorkerThread
-    fun markProcessed(key: Key?, processed: Boolean) {
-        key?.apply { db.translationDao().markProcessed(tool, language, version, processed) }
-    }
+internal abstract class TranslationRepository internal constructor(private val db: ArticleRoomDatabase) {
+    suspend fun isProcessed(translation: Translation) = find(translation.toTranslationRefKey())?.processed ?: false
+    private suspend fun find(key: Key?) = key?.run { db.translationDao().find(tool, language, version) }
 
     @Transaction
-    @WorkerThread
-    open fun addAemImports(translation: Translation, uris: List<Uri>): Boolean {
+    open suspend fun addAemImports(translation: Translation, uris: List<Uri>): Boolean {
         val translationKey = translation.toTranslationRefKey() ?: return false
 
         // create translation ref if it doesn't exist already
@@ -48,13 +34,18 @@ abstract class TranslationRepository internal constructor(private val db: Articl
         return true
     }
 
+    private suspend fun markProcessed(key: Key?, processed: Boolean) {
+        key?.apply { db.translationDao().markProcessed(tool, language, version, processed) }
+    }
+
     @Transaction
-    @WorkerThread
-    open fun removeMissingTranslations(translationsToKeep: List<Translation>) {
+    open suspend fun removeMissingTranslations(translationsToKeep: List<Translation>) {
+        val translationDao = db.translationDao()
         val valid = translationsToKeep.map { it.toTranslationRefKey() }.toSet()
-        db.translationDao().all
-            .filterNot { valid.contains(it.key) }
-            .apply { db.translationDao().remove(this) }
-        db.aemImportRepository().removeOrphanedAemImports()
+        val toRemove = translationDao.getAll().filterNot { valid.contains(it.key) }
+        if (toRemove.isNotEmpty()) {
+            translationDao.remove(toRemove)
+            db.aemImportRepository().removeOrphanedAemImports()
+        }
     }
 }
